Refactor: Moves functions around
Moves the sprite draw system into graphics. Moves movement code into its
own file.
